Create Index.html
List one is a list of items you can buy at a grocery store. There's a lot of different items, so it was a good list to pick for this type of assignment. 

List two ordered the items from least to most healthy. It's not perfect because different brands very in healthynes, but it's a numbered list, and is generally accurate.

The glossary is ordered in alphabetical order and gives a brief explanation of each of the items, and examples of broader items.